==> code/0.5_phys_parameters.Rmd <==
---
title: 'Hepatoprotective effects of systemic ER activation'
subtitle: 'Physiological parameters mice upon sacrifice'
author: 'Christian Sommerauer & Carlos Gallardo'
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  pdf_document:
    highlight: tango
---

```{r setup, include=FALSE}
# set path
path <- strsplit(dirname(rstudioapi::getActiveDocumentContext()$path), '/')[[1]]
path <- path[-c((length(path)-0):length(path))]
path <- paste(path, collapse = '/')
knitr::opts_knit$set(root.dir = path)
knitr::opts_chunk$set(echo = T, message = F, warning = F, dev = 'cairo_pdf')
def.chunk.hook  <- knitr::knit_hooks$get("chunk")
knitr::knit_hooks$set(chunk = function(x, options) {
  size <- 'scriptsize'
  x <- def.chunk.hook(x, options)
  paste0("\n \\", size, "\n\n", x, "\n\n \\normalsize")
})

```

```{r}
# library import
library(ggplot2)
library(dplyr)
```

```{r}
# color palettes
colPals <- list()
colPals$conditions <- setNames(c('#44AA99', '#117733', '#88CCEE', '#332288', '#DDCC77', '#CC6677', '#AA4499', '#882255'),
                               c('CDf', 'HFDf', 'CDm', 'HFDm', 'DPN', 'DIP', 'E2', 'PPT'))
colPals$RdBu <- rev(RColorBrewer::brewer.pal(n=11, name = 'RdBu'))
colPals$UpDown <- setNames(colPals$RdBu[c(10,2)],
                           c('up', 'down'))

```


# Load data

```{r}
phys_parameters <- read.delim("data/220923_physio_parameters.txt") %>% 
  dplyr::mutate(group = gsub('_fe','f',group)) %>% 
  dplyr::mutate(group = gsub('_ma','m',group)) %>% 
  dplyr::mutate(group = factor(group, levels = names(colPals$conditions)))

```


# Calculate means and standard deviations

```{r}
phys_para_avg_sd <- phys_parameters %>%
  group_by(group) %>%
  mutate(liver_average = mean(liver)) %>%
  mutate(glucose_average = mean(glucose)) %>%
  mutate(tot_weight_average = mean(total_weight)) %>%
  mutate(liver_sd = sd(liver)) %>%
  mutate(glucose_sd = sd(glucose)) %>%
  mutate(tot_weight_sd = sd(total_weight))

```

```{r}
# Plot the total weight
ggplot(phys_para_avg_sd, 
       aes(x=group,
           y = tot_weight_average)) +
  geom_errorbar(aes(ymin=tot_weight_average-tot_weight_sd, 
                    ymax=tot_weight_average+tot_weight_sd), 
                    width = 0.5, color="black") +
  geom_col(aes(fill=group), position= "dodge", color="black") +
  geom_point(aes(y = total_weight),shape=21,size=4, color="#5a6664", fill="#daf6f1", alpha=0.8) +
  theme_bw() +
  theme_classic()+
  theme(axis.text.x = element_text(angle=90, size = 20,  vjust = 0.5, hjust = 0.95),
        axis.text.y = element_text(size = 20),
        axis.title.y = element_text(size = 15),
        legend.position = "none") +
  scale_y_continuous(limits = c(0, 55)) +
  scale_fill_manual(values = colPals$condition) +
  xlab("") +
  ylab("average total weight in grams")


# Plot the liver weight
ggplot(phys_para_avg_sd, 
       aes(x=group,
           y = liver_average)) +
  geom_errorbar(aes(ymin=liver_average-liver_sd, 
                    ymax=liver_average+liver_sd), 
                    width = 0.5, color="black") +
  geom_col(aes(fill=group), position= "dodge", color="black") +
  geom_point(aes(y = liver),shape=21,size=4, color="#5a6664", fill="#daf6f1", alpha=0.8) +
  theme_bw() +
  theme_classic()+
  theme(axis.text.x = element_text(angle=90, size = 20,  vjust = 0.5, hjust = 0.95),
        axis.text.y = element_text(size = 20),
        axis.title.y = element_text(size = 15),
        legend.position = "none") +
  scale_y_continuous(limits = c(0, 2.5)) +
  scale_fill_manual(values = colPals$condition) +
  xlab("") +
  ylab("average liver weight in grams")


# Plot the blood glucose 2h before sacrifice 
ggplot(phys_para_avg_sd, 
       aes(x=group,
           y = glucose_average)) +
  geom_errorbar(aes(ymin=glucose_average-glucose_sd, 
                    ymax=glucose_average+glucose_sd), 
                    width = 0.5, color="black") +
  geom_col(aes(fill=group), position= "dodge", color="black") +
  geom_point(aes(y = glucose),shape=21,size=4, color="#5a6664", fill="#daf6f1", alpha=0.8) +
  theme_bw() +
  theme_classic()+
  theme(axis.text.x = element_text(angle=90, size = 20,  vjust = 0.5, hjust = 0.95),
        axis.text.y = element_text(size = 20),
        axis.title.y = element_text(size = 15),
        legend.position = "none") +
  scale_y_continuous(limits = c(0, 13)) +
  scale_fill_manual(values = colPals$condition) +
  xlab("") +
  ylab("average fasting glucose \n(2h) mmol/liter")

```


# Perform statistical tests

```{r}
total <- aov(phys_parameters$total_weight ~ phys_parameters$group)
TukeyHSD(total)

liver <- aov(phys_parameters$liver ~ phys_parameters$group)
TukeyHSD(liver)

glucose <- aov(phys_parameters$glucose ~ phys_parameters$group)
TukeyHSD(glucose)
```


# SessionInfo

```{r}
sessionInfo()

```
