==> code/04_single_cell_analysis.Rmd <==
---
title: 'Hepatoprotective effects of systemic ER activation'
subtitle: 'Single cell analysis'
author: 'Christian Sommerauer & Carlos Gallardo'
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  pdf_document:
    highlight: tango
---

```{r setup, include=FALSE}
# set path
path <- strsplit(dirname(rstudioapi::getActiveDocumentContext()$path), '/')[[1]]
path <- path[-c((length(path)-0):length(path))]
path <- paste(path, collapse = '/')
knitr::opts_knit$set(root.dir = path)
knitr::opts_chunk$set(echo = T, message = F, warning = F, dev = 'png')
def.chunk.hook  <- knitr::knit_hooks$get("chunk")
knitr::knit_hooks$set(chunk = function(x, options) {
  size <- 'scriptsize'
  x <- def.chunk.hook(x, options)
  paste0("\n \\", size, "\n\n", x, "\n\n \\normalsize")
})

```

```{r}
# source and library import
source('code/00_helper_functions.R')
library(tidyverse)
library(Seurat)
library(pagoda2)
library(patchwork)

```

```{r}
# color palettes
colPals <- list()
colPals$conditions <- setNames(c('#44AA99', '#117733', '#88CCEE', '#332288', '#DDCC77', '#CC6677', '#AA4499', '#882255'),
                               c('CDf', 'HFDf', 'CDm', 'HFDm', 'DPN', 'DIP', 'E2', 'PPT'))
colPals$RdBu <- rev(RColorBrewer::brewer.pal(n=11, name = 'RdBu'))
colPals$UpDown <- setNames(colPals$RdBu[c(10,2)],
                           c('up', 'down'))
colPals$clusters <- setNames(c('#E6E6E6', '#B3B3B3', '#8C8C8C', '#4D4D4D'),
                             c('1', '2', '3', '4'))
colPals$celltypes <- setNames(c('#B4272F', '#E5462D', '#FFD1D1', '#F4E54C', '#FBAA3E', '#AA654E', '#B58B80', '#6D7AA5', 
                                '#B3177E', '#CAC1DD', '#67227D','#36B449', '#82C349', '#A9D265', '#199478', '#95A3A3', '#C4C5C7'),
                              c('Cholangiocytes', 'Endothelial cells', 'HsPCs', 'Stromal cells', 'Hepatocytes', 
                                'Kupffer cells', 'Monocytes & Monocyte-derived cells', 'T cells', 'NK cells', 
                                'ILC1s', 'B cells', 'cDC1s', 'cDC2s', 'Mig. cDCs', 'pDCs', 'Basophils', 'Neutrophils'))
colPals$inferno <- c('#FCFFA4', '#FCA50A', '#DD513A', '#932667', '#420A68', '#000004')

```


# Note

This script is highly demanding in terms of computational resources. Approximately 50GB
of single-cell data from the Liver Cell Atlas resource are loaded into memory. Establishing
cell enrichments of gene signatures with `pagoda2` also requires long computing time. Run
using `pagoda2` v1.0.2 and `Seurat` v4.0.2 on R version 4.0.5


# Load data

```{r}
# Liver Cell Atlas data from Guilliams et al. 2022 (https://www.livercellatlas.org/download.php)
liver <- readRDS(file = 'data/livercellatlas_feb2022.rds')

# Liver Cell Atlas sample metadata
liver_meta <- read.table(
  file = 'data/livercellatlas_feb2022_sample_annotation.tsv',
  stringsAsFactors = FALSE,
  sep = '\t',
  header = TRUE)

# gene sets
DEG_sets <- readRDS('results/bulkRNAseq_mmus_DEG_sets.rds')
pathway_sets <- readRDS('results/bulkRNAseq_mmus_GSEA_reactome_cluster_sets.rds')

# mouse-human orthologs
mouse_human_orthologs <- read.table(
  file = 'data/ensembl_mmus_hsap_sep2019_orthologs.tsv',
  sep = '\t',
  header = TRUE,
  quote = '')

# mouse-macaque orthologs
mouse_macaque_orthologs <- read.table(
  file = 'data/ensembl_mmus_mfas_sep2019_orthologs.tsv',
  sep = '\t',
  header = TRUE,
  quote = '')

```


# Mouse single-cell map

```{r}
# filter for male-only samples and remove HsPCs
df <- liver$mouseStSt$All@meta.data %>% 
  dplyr::left_join(liver_meta, by = 'sample') %>% 
  dplyr::filter(sex == 'male' & annot != 'HsPCs') %>% 
  dplyr::mutate(annot = replace(annot, annot == 'Fibroblasts', 'Stromal cells'))

ggplot(df, aes(x=UMAP_1, y=UMAP_2, color=annot, fill=annot, label=annot)) +
  geom_point(shape=21, size=1) +
  scale_fill_manual(values = colPals$celltypes) +
  scale_color_manual(values = colPals$celltypes) +
  xlab("UMAP 1") + ylab("UMAP 2") +
  theme_bw() +
  guides(colour = guide_legend(override.aes = list(size=3))) +
  theme(legend.title = element_blank())

```


# Cell type composition

```{r}
# filter for male-only samples and remove HsPCs
# filter for cells from in vivo digestion protocol
df <- liver$mouseStSt$All@meta.data %>% 
  dplyr::left_join(liver_meta, by = 'sample') %>% 
  dplyr::filter(sex == 'male' & annot != 'HsPCs') %>% 
  dplyr::mutate(annot = replace(annot, annot == 'Fibroblasts', 'Stromal cells')) %>% 
  dplyr::filter(digest == 'inVivo') %>% 
  dplyr::group_by(annot) %>% 
  dplyr::summarize(mouseStSt_n = dplyr::n()) %>% 
  dplyr::mutate(mouseStSt_freq = round(mouseStSt_n / sum(mouseStSt_n) * 100, 2))

df2 <- liver$mouseNAFLD$All@meta.data %>% 
  dplyr::left_join(liver_meta, by = 'sample') %>% 
  dplyr::filter(sex == 'male' & annot != 'HsPCs') %>% 
  dplyr::filter(digest == 'inVivo') %>% 
  dplyr::mutate(annot = replace(annot, annot == 'KCs', 'Kupffer cells')) %>% 
  dplyr::mutate(annot = replace(annot, annot == 'Mono/mono-derived cells', 'Monocytes & Monocyte-derived cells')) %>% 
  dplyr::mutate(annot = replace(annot, annot == 'Mig cDCs', 'Mig. cDCs')) %>% 
  dplyr::group_by(annot) %>% 
  dplyr::summarize(mouseNAFLD_n = dplyr::n()) %>% 
  dplyr::mutate(mouseNAFLD_freq = round(mouseNAFLD_n / sum(mouseNAFLD_n) * 100, 2))

df <- dplyr::full_join(df, df2, by = 'annot') %>% 
  replace(is.na(.), 0) %>% 
  dplyr::mutate(mouseStSt_minus_NAFLD_freq = mouseStSt_freq - mouseNAFLD_freq) %>% 
  tidyr::pivot_longer(cols = c(mouseStSt_freq, mouseNAFLD_freq), names_to = 'cond', values_to = 'freq') %>% 
  dplyr::mutate(annot = factor(annot, levels = rev(names(colPals$celltypes))),
                cond = factor(cond, levels = c('mouseStSt_freq','mouseNAFLD_freq')))

ggplot(df, aes(x=freq, y=annot, fill=cond, label=freq)) +
  geom_bar(color="black", size=0.5, width=0.8, position="stack", stat="identity") +
  geom_text(size=3, hjust=-0.2) +
  scale_fill_manual(values = unname(colPals$conditions[c('CDm', 'HFDm')])) +
  scale_x_continuous(expand = expansion(mult = c(.0, .1)),
                     limit = c(0, 50),
                     breaks = c(0, 25, 50)) +
  facet_wrap(~cond) +
  xlab('Abundance (%)') +
  theme_bw() +
  theme(strip.background = element_blank())

```

```{r}
# cell type frequency change (%CD - %HFD)
df2 <- df %>% 
  dplyr::select(annot, mouseStSt_minus_NAFLD_freq) %>% 
  dplyr::filter(!duplicated(.)) %>% 
  dplyr::mutate(change=ifelse(mouseStSt_minus_NAFLD_freq>0, 'up', 'down'))

ggplot(df2, aes(x=mouseStSt_minus_NAFLD_freq, y=annot, fill=change)) +
  geom_bar(color="black", size=0.5, width=0.8, position="stack", stat="identity") +
  scale_fill_manual(values = colPals$UpDown) +
  scale_x_continuous(expand = expansion(mult = c(.0, .1)),
                     limit = c(-25, 25),
                     breaks = c(-20, 0, 20)) +
  xlab('Change (%)') +
  theme_bw() +
  theme(legend.position = 'none')

```


# Cell type specificity of ER activation signatures

```{r, fig.width=12, fig.height=3.5}
# subsample single-cell data to 5000 cells per cell type for enrichment analysis
set.seed(22)
dat <- liver$mouseStSt$All
subs <- dat@meta.data %>% 
  tibble::rownames_to_column(var = 'cellid') %>% 
  dplyr::left_join(liver_meta, by = 'sample') %>% 
  dplyr::filter(sex == 'male' & annot != 'HsPCs') %>% 
  dplyr::group_by(annot) %>% 
  dplyr::slice_sample(n=5000)
dat <- subset(dat, cells = subs$cellid)

res <- getPAS(x = dat, 
              gene.sets = DEG_sets$gene_symbols, 
              npcs = 100)

p <- lapply(names(DEG_sets$gene_symbols), function(set) {
  local({
    df <- data.frame(UMAP_1 = res$UMAP_1,
                     UMAP_2 = res$UMAP_2)
    df$score <- res[, set]
    p1 <- ggplot(df, aes(x=UMAP_1, y=UMAP_2, color=score)) +
      geom_point(shape=19, size=1) +
      scale_color_gradientn(colours = colPals$inferno) +
      theme_void() +
      theme(legend.position = 'none',
            plot.title = element_text(hjust = 0.5)) +
      ggtitle(set)
  })
})

patchwork::wrap_plots(p, nrow=1, ncol=4, byrow=T, guides = 'collect')

```


# Conservation of cell type specificities in primates

## Human

```{r, fig.width=12, fig.height=3.5}
# subsample single-cell data to 5000 cells per cell type for enrichment analysis
set.seed(22)
dat <- liver$human$All
subs <- dat@meta.data %>% 
  tibble::rownames_to_column(var = 'cellid') %>% 
  dplyr::left_join(liver_meta, by = 'sample') %>% 
  dplyr::filter(sex == 'male') %>% 
  dplyr::group_by(annot) %>% 
  dplyr::slice_sample(n=5000)
dat <- subset(dat, cells = subs$cellid)

DEG_sets_human <- lapply(DEG_sets$gene_id, function(x) {
  mouse_human_orthologs %>% 
    dplyr::filter(GeneID_mouse %in% x) %>% 
    dplyr::pull(GeneSymbol_human) %>% 
    unique()
})

res <- getPAS(x = dat, 
              gene.sets = DEG_sets_human, 
              npcs = 100)

p <- lapply(names(DEG_sets_human), function(set) {
  local({
    df <- data.frame(UMAP_1 = res$UMAP_1,
                     UMAP_2 = res$UMAP_2)
    df$score <- res[, set]
    p1 <- ggplot(df, aes(x=UMAP_1, y=UMAP_2, color=score)) +
      geom_point(shape=19, size=1) +
      scale_color_gradientn(colours = colPals$inferno) +
      theme_void() +
      theme(legend.position = 'none',
            plot.title = element_text(hjust = 0.5)) +
      ggtitle(set)
  })
})

patchwork::wrap_plots(p, nrow=1, ncol=4, byrow=T, guides = 'collect')

```


## Macaque

```{r, fig.width=12, fig.height=3.5}
# subsample single-cell data to 5000 cells per cell type for enrichment analysis
set.seed(22)
dat <- liver$multisp$monkey
subs <- dat@meta.data %>% 
  tibble::rownames_to_column(var = 'cellid') %>% 
  dplyr::left_join(liver_meta, by = 'sample') %>% 
  dplyr::filter(sex == 'male') %>% 
  dplyr::group_by(annot) %>% 
  dplyr::slice_sample(n=5000)
dat <- subset(dat, cells = subs$cellid)

DEG_sets_macaque <- lapply(DEG_sets$gene_id, function(x) {
  mouse_macaque_orthologs %>% 
    dplyr::filter(GeneID_mouse %in% x) %>% 
    dplyr::pull(GeneSymbol_macaque) %>% 
    unique()
})

res <- getPAS(x = dat, 
              gene.sets = DEG_sets_macaque, 
              npcs = 100)

p <- lapply(names(DEG_sets_macaque), function(set) {
  local({
    df <- data.frame(UMAP_1 = res$UMAP_1,
                     UMAP_2 = res$UMAP_2)
    df$score <- res[, set]
    p1 <- ggplot(df, aes(x=UMAP_1, y=UMAP_2, color=score)) +
      geom_point(shape=19, size=1) +
      scale_color_gradientn(colours = colPals$inferno) +
      theme_void() +
      theme(legend.position = 'none',
            plot.title = element_text(hjust = 0.5)) +
      ggtitle(set)
  })
})

patchwork::wrap_plots(p, nrow=1, ncol=4, byrow=T, guides = 'collect')

```


# Liver zoonation of ER activation signatures (spatial transcriptomics)

```{r, fig.width=12, fig.height=3.5}
# subsample single-cell data to 5000 cells per cell type for enrichment analysis
set.seed(22)
dat <- liver$mouseStSt$Visium
subs <- dat@meta.data %>% 
  tibble::rownames_to_column(var = 'cellid') %>% 
  dplyr::left_join(liver_meta, by = 'sample') %>% 
  dplyr::filter(sex == 'male') %>% 
  dplyr::group_by(zonationGroup) %>% 
  dplyr::slice_sample(n=5000)
dat <- subset(dat, cells = subs$cellid)

res <- getPAS(x = dat, 
              gene.sets = DEG_sets$gene_symbols, 
              npcs = 100)

p <- lapply(names(DEG_sets$gene_symbols), function(set) {
  local({
    df <- data.frame(UMAP_1 = res$UMAP_1,
                     UMAP_2 = res$UMAP_2)
    df$score <- res[, set]
    p1 <- ggplot(df, aes(x=UMAP_1, y=UMAP_2, color=score)) +
      geom_point(shape=19, size=1) +
      scale_color_gradientn(colours = colPals$inferno) +
      theme_void() +
      theme(legend.position = 'none',
            plot.title = element_text(hjust = 0.5)) +
      ggtitle(set)
  })
})

patchwork::wrap_plots(p, nrow=1, ncol=4, byrow=T, guides = 'collect')

```


# Pathway enrichment across cell types

## Steady State

```{r, fig.width=12, fig.height=10}
# subsample single-cell data to 5000 cells per cell type for enrichment analysis
set.seed(22)
dat <- liver$mouseStSt$All
subs <- dat@meta.data %>% 
  tibble::rownames_to_column(var = 'cellid') %>% 
  dplyr::left_join(liver_meta, by = 'sample') %>% 
  dplyr::filter(sex == 'male' & annot != 'HsPCs') %>% 
  dplyr::group_by(annot) %>% 
  dplyr::slice_sample(n=5000)
dat <- subset(dat, cells = subs$cellid)

res <- getPAS(x = dat, 
              gene.sets = pathway_sets, 
              npcs = 100)

df <- res %>% 
  dplyr::mutate(annot = replace(annot, annot == 'Fibroblasts', 'Stromal cells')) %>% 
  dplyr::group_by(annot) %>% 
  dplyr::summarize(dplyr::across(`Adaptive immune system`:`Vitamin metabolism`, mean)) %>% 
  dplyr::mutate(across(`Adaptive immune system`:`Vitamin metabolism`, ~ (.-min(.)) / (max(.)-min(.)) )) %>% 
  tidyr::pivot_longer(cols = `Adaptive immune system`:`Vitamin metabolism`, names_to = 'pathway', values_to = 'score') %>% 
  dplyr::mutate(annot = factor(annot, levels = rev(names(colPals$celltypes))),
                pathway = factor(pathway, levels = names(pathway_sets)))

p <- lapply(names(pathway_sets), function(set) {
  local({
    df <- data.frame(UMAP_1 = res$UMAP_1,
                     UMAP_2 = res$UMAP_2)
    df$score <- res[, set]
    p1 <- ggplot(df, aes(x=UMAP_1, y=UMAP_2, color=score)) +
      geom_point(shape=19, size=1) +
      scale_color_gradientn(colours = colPals$inferno) +
      theme_void() +
      theme(legend.position = 'none',
            plot.title = element_text(hjust = 0.5)) +
      ggtitle(set)
  })
})

patchwork::wrap_plots(p, ncol=5, byrow=T, guides = 'collect')

```


## NAFLD

```{r, fig.width=12, fig.height=10}
# subsample single-cell data to 5000 cells per cell type for enrichment analysis
set.seed(22)
dat <- liver$mouseNAFLD$All
subs <- dat@meta.data %>% 
  tibble::rownames_to_column(var = 'cellid') %>% 
  dplyr::left_join(liver_meta, by = 'sample') %>% 
  dplyr::filter(sex == 'male' & annot != 'HsPCs') %>% 
  dplyr::group_by(annot) %>% 
  dplyr::slice_sample(n=5000)
dat <- subset(dat, cells = subs$cellid)

res <- getPAS(x = dat, 
              gene.sets = pathway_sets, 
              npcs = 100)

p <- lapply(names(pathway_sets), function(set) {
  local({
    df <- data.frame(UMAP_1 = res$UMAP_1,
                     UMAP_2 = res$UMAP_2)
    df$score <- res[, set]
    p1 <- ggplot(df, aes(x=UMAP_1, y=UMAP_2, color=score)) +
      geom_point(shape=19, size=1) +
      scale_color_gradientn(colours = colPals$inferno) +
      theme_void() +
      theme(legend.position = 'none',
            plot.title = element_text(hjust = 0.5)) +
      ggtitle(set)
  })
})

patchwork::wrap_plots(p, ncol=5, byrow=T, guides = 'collect')

```


## Comparison of steady state vs NAFLD

```{r, fig.height = 12, fig.width = 18}
df2 <- res %>% 
  dplyr::mutate(annot = replace(annot, annot == 'KCs', 'Kupffer cells')) %>% 
  dplyr::mutate(annot = replace(annot, annot == 'Mono/mono-derived cells', 'Monocytes & Monocyte-derived cells')) %>% 
  dplyr::mutate(annot = replace(annot, annot == 'Mig cDCs', 'Mig. cDCs')) %>%
  dplyr::group_by(annot) %>% 
  dplyr::summarize(dplyr::across(`Adaptive immune system`:`Vitamin metabolism`, mean)) %>% 
  dplyr::mutate(dplyr::across(`Adaptive immune system`:`Vitamin metabolism`, ~ (.-min(.)) / (max(.)-min(.)) )) %>% 
  tidyr::pivot_longer(cols = `Adaptive immune system`:`Vitamin metabolism`, names_to = 'pathway', values_to = 'score') %>% 
  dplyr::mutate(cond = 'HFD')

df3 <- df %>% 
  dplyr::mutate(cond = 'CD') %>% 
  dplyr::bind_rows(df2) %>% 
  dplyr::mutate(cond = factor(cond, levels = c('CD','HFD')),
                annot = factor(annot, levels = names(colPals$celltypes)),
                pathway = factor(pathway, levels = rev(names(pathway_sets))))

ggplot(df3, aes(x=annot, y=pathway, size=score, color=score)) +
  geom_point() +
  scale_color_gradientn(colours = colPals$inferno) +
  scale_size(range = c(1,8))+
  facet_wrap(~cond) +
  theme_bw() +
  theme(strip.background = element_blank(),
        axis.text.x = element_text(angle = 45, hjust = 1, vjust = 1),
        axis.text.y = element_text(hjust = 1, vjust = 0.3))

```

```{r, fig.height = 12, fig.width = 11}
# Cell type enrichment change (CD - HFD)
df4 <- df %>% 
  dplyr::mutate(score = score - df2$score) %>% 
  dplyr::mutate(size = abs(score)) %>% 
  dplyr::mutate(annot = factor(annot, levels = names(colPals$celltypes)),
                pathway = factor(pathway, levels = rev(names(pathway_sets))))

ggplot(df4, aes(x=annot, y=pathway, size=size, color=score)) +
  geom_point() +
  scale_color_gradientn(colours = colPals$RdBu) +
  scale_size(range = c(1,8), breaks = c(0,0.5,1), limits = c(0,1))+
  theme_bw() +
  theme(strip.background = element_blank(),
        axis.text.x = element_text(angle = 45, hjust = 1, vjust = 1),
        axis.text.y = element_text(hjust = 1, vjust = 0.3))

```


# Pathway zonation in the liver

```{r, fig.width=12, fig.height=10}
# subsample single-cell data to 5000 cells per cell type for enrichment analysis
set.seed(22)
dat <- liver$mouseStSt$Visium
subs <- dat@meta.data %>% 
  tibble::rownames_to_column(var = 'cellid') %>% 
  dplyr::left_join(liver_meta, by = 'sample') %>% 
  dplyr::filter(sex == 'male') %>% 
  dplyr::group_by(zonationGroup) %>% 
  dplyr::slice_sample(n=5000)
dat <- subset(dat, cells = subs$cellid)

res <- getPAS(x = dat, 
              gene.sets = pathway_sets, 
              npcs = 100)

p <- lapply(names(pathway_sets), function(set) {
  local({
    df <- data.frame(UMAP_1 = res$UMAP_1,
                     UMAP_2 = res$UMAP_2)
    df$score <- res[, set]
    p1 <- ggplot(df, aes(x=UMAP_1, y=UMAP_2, color=score)) +
      geom_point(shape=19, size=1) +
      scale_color_gradientn(colours = colPals$inferno) +
      theme_void() +
      theme(legend.position = 'none',
            plot.title = element_text(hjust = 0.5)) +
      ggtitle(set)
  })
})

patchwork::wrap_plots(p, ncol=5, byrow=T, guides = 'collect')

```


# SessionInfo

```{r}
sessionInfo()

```
