==> code/06_candidate_identification.Rmd <==
---
title: 'Hepatoprotective effects of systemic ER activation'
subtitle: 'Human ER-sensitive genes and NAFLD classification models'
author: 'Christian Sommerauer & Carlos Gallardo'
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  pdf_document:
    highlight: tango
---

```{r setup, include=FALSE}
# set path
path <- strsplit(dirname(rstudioapi::getActiveDocumentContext()$path), '/')[[1]]
path <- path[-c((length(path)-0):length(path))]
path <- paste(path, collapse = '/')
knitr::opts_knit$set(root.dir = path)
knitr::opts_chunk$set(echo = T, message = F, warning = F, dev = 'cairo_pdf')
def.chunk.hook  <- knitr::knit_hooks$get("chunk")
knitr::knit_hooks$set(chunk = function(x, options) {
  size <- 'scriptsize'
  x <- def.chunk.hook(x, options)
  paste0("\n \\", size, "\n\n", x, "\n\n \\normalsize")
})

```


```{r}
# source and library import
source('code/00_helper_functions.R')
library(tidyverse)
library(RColorBrewer)
library(ComplexHeatmap)
library(patchwork)

```

```{r}
# color palettes
colPals <- list()
colPals$conditions <- setNames(c('#44AA99', '#117733', '#88CCEE', '#332288', '#DDCC77', '#CC6677', '#AA4499', '#882255'),
                               c('CDf', 'HFDf', 'CDm', 'HFDm', 'DPN', 'DIP', 'E2', 'PPT'))
colPals$RdBu <- rev(RColorBrewer::brewer.pal(n=11, name = 'RdBu'))
colPals$UpDown <- setNames(colPals$RdBu[c(10,2)],
                           c('up', 'down'))
colPals$clusters <- setNames(c('#E6E6E6', '#B3B3B3', '#8C8C8C', '#4D4D4D'),
                             c('1', '2', '3', '4'))
colPals$celltypes <- setNames(c('#B4272F', '#E5462D', '#FFD1D1', '#F4E54C', '#FBAA3E', '#AA654E', '#B58B80', '#6D7AA5', 
                                '#B3177E', '#CAC1DD', '#67227D','#36B449', '#82C349', '#A9D265', '#199478', '#95A3A3', '#C4C5C7'),
                              c('Cholangiocytes', 'Endothelial cells', 'HsPCs', 'Stromal cells', 'Hepatocytes', 
                                'Kupffer cells', 'Monocytes & Monocyte-derived cells', 'T cells', 'NK cells', 
                                'ILC1s', 'B cells', 'cDC1s', 'cDC2s', 'Mig. cDCs', 'pDCs', 'Basophils', 'Neutrophils'))
colPals$inferno <- c('#FCFFA4', '#FCA50A', '#DD513A', '#932667', '#420A68', '#000004')
colPals$models <- setNames(c('#F6B651', '#92A3A5', '#972D22'),
                           c('stage', 'nas', 'fibrosis'))
colPals$models_light <- setNames(c('#FAD7A0', '#BFC9CA', '#DF7B71'),
                                 c('stage', 'nas', 'fibrosis'))

```


# Load data

```{r}
# mouse estrogen-sensitive enhancer-gene pairs
mouse_49_ER_genes <- read.table(
  file = 'results/Epigenome_analysis/corr_49genes_76enh_toPlot.txt',
  stringsAsFactors = FALSE,
  sep = '\t',
  header = TRUE) %>% 
  dplyr::pull(symbol) %>% 
  unique()

# mouse-human orthologs
mouse_human_orthologs <- read.table(
  file = 'data/ensembl_mmus_hsap_dec2021_orthologs.tsv',
  sep = '\t',
  header = TRUE,
  quote = '')

# mouse differentially expressed genes
DEGs <- readRDS('results/bulkRNAseq_mmus_DEGs.rds')

# NAFLD patient cohort
cohort_data <- readRDS("data/bulkRNAseq_human_cohort_data.rds")

cohort_data$Govaere$cpm_filt <- cohort_data$Govaere$cpm %>%
  tibble::rownames_to_column(var = 'gene') %>%
  dplyr::filter(gene %in% mouse_human_orthologs$GeneID_human) %>%
  dplyr::mutate(gene = dplyr::recode(gene, !!!setNames(mouse_human_orthologs$GeneSymbol_human,
                                                       mouse_human_orthologs$GeneID_human))) %>%
  dplyr::filter(!duplicated(gene) & gene != "") %>%
  tibble::column_to_rownames(var = 'gene')

```


# Convert estrogen-sensitive genes in mouse to human orthologs

```{r}
# find genes that are expressed in human patients. Cutoff: 0.5 CPM.
human_detectable <- cohort_data$Govaere$cpm_filt %>% 
  apply(., 1, median)
human_detectable <- names(human_detectable[human_detectable > 0.5])

# filter ER genes for those expressed in human
human_ER_genes <- mouse_human_orthologs %>% 
  dplyr::filter(GeneSymbol_mouse %in% mouse_49_ER_genes) %>% 
  dplyr::filter(GeneSymbol_human %in% human_detectable) %>%
  dplyr::filter(!duplicated(GeneSymbol_human)) %>%
  dplyr::pull(GeneSymbol_human)

# export human ER genes
cat(human_ER_genes,
    file = 'results/ER_sensitive_genes.txt',
    sep = '\n')

```


# Integrated heatmap of ER-sensitive genes across human NAFLD spectrum

```{r, fig.width=12, fig.height=10}
human_stage <- cohort_data$Govaere$cpm_filt %>% 
  dplyr::filter(rownames(.) %in% human_ER_genes) %>% 
  groupTransform(cohort_data$Govaere$meta$Stage, function(x) apply(x,1,median)) %>% 
  scaleData(method = 'zscore') %>% 
  dplyr::arrange(rownames(.))

human_steatosis <- cohort_data$Govaere$cpm_filt %>% 
  dplyr::filter(rownames(.) %in% human_ER_genes) %>% 
  groupTransform(cohort_data$Govaere$meta$NAS, function(x) apply(x,1,median)) %>% 
  scaleData(method = 'zscore') %>% 
  dplyr::arrange(rownames(.))

human_fibrosis <- cohort_data$Govaere$cpm_filt %>% 
  dplyr::filter(rownames(.) %in% human_ER_genes) %>% 
  groupTransform(cohort_data$Govaere$meta$Fibrosis, function(x) apply(x,1,median)) %>% 
  scaleData(method = 'zscore') %>% 
  dplyr::arrange(rownames(.))

mouse_log2FC_HFDmVsCDm <- DEGs$filt$CDmVsHFDm %>% 
  dplyr::rename(GeneSymbol_mouse = external_gene_name) %>% 
  dplyr::inner_join(mouse_human_orthologs, by = 'GeneSymbol_mouse') %>% 
  dplyr::filter(GeneSymbol_human %in% human_ER_genes) %>% 
  dplyr::filter(!duplicated(GeneSymbol_human)) %>% 
  dplyr::select(GeneSymbol_human, log2FoldChange) %>% 
  dplyr::mutate(log2FoldChange = log2FoldChange*-1) %>% 
  dplyr::rename(HFDmVsCDm = log2FoldChange) %>% 
  dplyr::mutate(HFDmVsCDm = factor(ifelse(HFDmVsCDm>0, 'high', 'low'), levels = c('high','low'))) %>% 
  tibble::column_to_rownames(var = 'GeneSymbol_human') %>% 
  dplyr::arrange(rownames(.))

mouse_log2FC_HFDfVsCDf <- DEGs$filt$CDfVsHFDf %>% 
  dplyr::rename(GeneSymbol_mouse = external_gene_name) %>% 
  dplyr::inner_join(mouse_human_orthologs, by = 'GeneSymbol_mouse') %>% 
  dplyr::filter(GeneSymbol_human %in% human_ER_genes) %>% 
  dplyr::filter(!duplicated(GeneSymbol_human)) %>% 
  dplyr::select(GeneSymbol_human, log2FoldChange) %>% 
  dplyr::mutate(log2FoldChange = log2FoldChange*-1) %>% 
  dplyr::rename(HFDfVsCDf = log2FoldChange) %>% 
  dplyr::mutate(HFDfVsCDf = ifelse(HFDfVsCDf>0, 'high', 'low')) %>% 
  dplyr::add_row(GeneSymbol_human = dplyr::setdiff(rownames(mouse_log2FC_HFDmVsCDm),
                                                   .$GeneSymbol_human), 
                 HFDfVsCDf = 'unchanged') %>% 
  dplyr::mutate(HFDfVsCDf = factor(HFDfVsCDf, levels = c('high','low', 'unchanged'))) %>% 
  tibble::column_to_rownames(var = 'GeneSymbol_human') %>% 
  dplyr::arrange(rownames(.))

set.seed(30)
clusters_ER_genes <- kmeans(human_stage, centers = 4, iter.max = 100)

Heatmap(human_stage, width = unit(24, "mm"),name = "Stage", 
        split = clusters_ER_genes$cluster,
        col = circlize::colorRamp2(breaks=seq(-max(abs(human_stage)), max(abs(human_stage)), length.out=11),
                                   colors=colPals$RdBu),
        row_title = c("Cluster 1", "Cluster 2", "Cluster 3", "Cluster 4"),
        cluster_row_slices = FALSE, 
        column_order=c("CTRL", "NAFL", "NASH")) +
  Heatmap(human_steatosis, width = unit(72, "mm"),name = "Steatosis",
          column_order=paste0('NAS', seq(0,8)),
          col = circlize::colorRamp2(breaks=seq(-max(abs(human_steatosis)), max(abs(human_steatosis)), length.out=11),
                                     colors=colPals$RdBu)) +
  Heatmap(human_fibrosis, width = unit(40, "mm"),name = "Fibrosis",
          column_order=paste0('F', seq(0,4)),
          col = circlize::colorRamp2(breaks=seq(-max(abs(human_fibrosis)), max(abs(human_fibrosis)), length.out=11),
                                     colors=colPals$RdBu)) +
  Heatmap(mouse_log2FC_HFDmVsCDm, width = unit(8, "mm"), name = "Male mouse log2FC",
          col = c('#199478', '#67227D')) +
  Heatmap(mouse_log2FC_HFDfVsCDf, width = unit(8, "mm"), name = "Female mouse log2FC",
          col = c('#F29724', '#8C510A', 'grey'))

```


# SessionInfo

```{r}
sessionInfo()

```
